Bolt-On Joint Tubes Marked as "Spec" in MFG
Vertical tubes that have a bolt-on, when exported, get "-spec" appended to their description. This keeps them from being merged together with others of the same length (which are likely different anyways), and will hopefully let the engineer know that it requires more work to be done for it.

No fancy tube-standard-identification is done yet.